+++ src/teamster/kipptaf/resources.py
@@ -10,7 +10,7 @@ from .alchemer.resources import AlchemerResource
 from .amplify.resources import MClassResource
 from .google.directory.resources import GoogleDirectoryResource
 from .google.forms.resources import GoogleFormsResource
-from .google.sheets.resources import GoogleSheetsResource
+from .google.resources import GoogleSheetsResource
 from .ldap.resources import LdapResource
 from .schoolmint.grow.resources import SchoolMintGrowResource
 from .smartrecruiters.resources import SmartRecruitersResource
